Introduce style::values::animated::color
Starting to remove stuff from style::properties::animated_properties for real.
