+++ components/style/properties/longhand/svg.mako.rs
@@ -20,11 +20,13 @@ ${helpers.single_keyword("vector-effect", "none non-scaling-stroke",
 // Section 13 - Gradients and Patterns
 
 ${helpers.predefined_type(
-    "stop-color", "RGBAColor",
+    "stop-color",
+    "RGBAColor",
     "RGBA::new(0, 0, 0, 255)",
     products="gecko",
-    animation_value_type="IntermediateRGBA",
-    spec="https://www.w3.org/TR/SVGTiny12/painting.html#StopColorProperty")}
+    animation_value_type="AnimatedRGBA",
+    spec="https://www.w3.org/TR/SVGTiny12/painting.html#StopColorProperty",
+)}
 
 ${helpers.predefined_type("stop-opacity", "Opacity", "1.0",
                           products="gecko",
@@ -34,22 +36,26 @@ ${helpers.predefined_type("stop-opacity", "Opacity", "1.0",
 // Section 15 - Filter Effects
 
 ${helpers.predefined_type(
-    "flood-color", "RGBAColor",
+    "flood-color",
+    "RGBAColor",
     "RGBA::new(0, 0, 0, 255)",
     products="gecko",
-    animation_value_type="IntermediateRGBA",
-    spec="https://www.w3.org/TR/SVG/filters.html#FloodColorProperty")}
+    animation_value_type="AnimatedRGBA",
+    spec="https://www.w3.org/TR/SVG/filters.html#FloodColorProperty",
+)}
 
 ${helpers.predefined_type("flood-opacity", "Opacity",
                           "1.0", products="gecko", animation_value_type="ComputedValue",
                           spec="https://www.w3.org/TR/SVG/filters.html#FloodOpacityProperty")}
 
 ${helpers.predefined_type(
-    "lighting-color", "RGBAColor",
+    "lighting-color",
+    "RGBAColor",
     "RGBA::new(255, 255, 255, 255)",
     products="gecko",
-    animation_value_type="IntermediateRGBA",
-    spec="https://www.w3.org/TR/SVG/filters.html#LightingColorProperty")}
+    animation_value_type="AnimatedRGBA",
+    spec="https://www.w3.org/TR/SVG/filters.html#LightingColorProperty",
+)}
 
 // CSS Masking Module Level 1
 // https://drafts.fxtf.org/css-masking
